Cantina audit fixes (#186)
* fix: round shares up on withdraw

* fix: use uint120s for timestamps in auction

* docs: natspec updates explaining risk profile of leverage tokens

* fix: remove decimals offset and align decimals on 18 (#179)

   - removed decimals offset entirely and added special if statement to handle edge case when strategy is empty
   - removed +1 from convertToShares function because if statement is not handling zero division and multiplication
   - added alignment on 18 decimals when initial shares are minted to the user so now no matter on how many decimals is collateral asset leverage token will always have 18 decimals precision

* fix: strict collateral ratio check (#184)

* feat: add re-entrancy guard to createNewLeverageToken, deposit, withdraw, and rebalance

---------

+++ test/unit/DutchAuctionRebalanceAdapter/Initialize.t.sol
@@ -12,7 +12,7 @@ import {DutchAuctionRebalanceAdapterTest} from "./DutchAuctionRebalanceAdapter.t
 import {DutchAuctionRebalanceAdapterHarness} from "test/unit/harness/DutchAuctionRebalanceAdapterHarness.t.sol";
 
 contract InitializeTest is DutchAuctionRebalanceAdapterTest {
-    function test_Fuzz_Initialize(uint256 auctionDuration, uint256 initialPriceMultiplier, uint256 minPriceMultiplier)
+    function test_Fuzz_Initialize(uint120 auctionDuration, uint256 initialPriceMultiplier, uint256 minPriceMultiplier)
         public
     {
         vm.assume(auctionDuration > 0);
@@ -54,7 +54,7 @@ contract InitializeTest is DutchAuctionRebalanceAdapterTest {
     }
 
     function test_Fuzz_Initialize_RevertIf_MinPriceMultiplierTooHigh(
-        uint256 auctionDuration,
+        uint120 auctionDuration,
         uint256 initialPriceMultiplier,
         uint256 minPriceMultiplier
     ) public {
